chore: split up auth migration test
